@@ -9,18 +9,18 @@
     <packaging>jar</packaging>
 
     <name>spring-boot-docker</name>
-    <description>Spring Boot Docker configuration study</description>
-
-    <parent>
+    <description>Spring Boot Docker configuration study</description>    <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>1.4.3.RELEASE</version> <!-- lookup parent from repository -->
+        <version>3.0.0</version>
+        <relativePath/> <!-- lookup parent from repository -->
     </parent>
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <java.version>1.7</java.version>
+        <java.version>17</java.version>
+        <testcontainers.version>1.18.3</testcontainers.version>
     </properties>
 
     <profiles>
